Separate stress tests to another target
